fix: move the billing banner into the workspace settings panel
The prototype on comfydesigner/team-workspaces-v1 mounts BillingStatusBanner
only inside the settings panels (MembersPanelContent, PlanCreditsPanelContent,
PartnerNodesPanelContent) and puts no billing banner in GraphView or LinearView.
Mounting it in the app shell was unfounded — match the prototype and Figma.

Ours takes a single slot in WorkspacePanelContent between the tab list and the
tab body, which is what "one banner above every settings tab" means for our
two-tab container, rather than duplicating it into each panel.

Drops the app-shell wrapper padding: spacing now comes from the mount site, and
the @container stays on the banner so it isn't coupled to a parent's classes.

The GraphView/LinearView placement tests go with the mounts. WorkspacePanelContent
gains one pinning the banner above the tab body.
